add changes to allow full subclassing of Applify
It is *almost* possible to subclass Applify, which would allow for
more customization than is currently possible.  The only remaining
changes required are to the installation of _app_run() and _app_new()
into the application class.  The new code installs wrappers which
ensure that the methods used follow the class inheritance.